feat: migrate service routes to App Router (Phase 4)
- Add src/app/services/page.tsx as Server Component (services hub)
- Add src/app/services/[...slug]/page.tsx as Server Component with catch-all
- Replace getStaticPaths with generateStaticParams, getStaticProps with inline data fetching
- Export metadata and generateMetadata for dynamic SEO
- Move all JSON-LD (CollectionPage, Service, LocalBusiness, FAQ, Breadcrumb) to inline <script> tags
- Use SanityImage component instead of Image+loader to avoid function serialization across server/client boundary
- Delete src/pages/services/ directory
